[stm32-ocm3] Expand clock module implementation
In certain cases, delaying may be off by up to an entire uSec, when
the systick counter is about to wrap.  The implementation here is not the
most precise, with perfectly counted assembly loops, but it's clear and
functional.  You shouldn't be using these low level routines for anything
critical anyway.