cpu/arm/stm32-ocm3/clock.c: more efficient clock_delay_usec; tighten calculations
This commit picks up the original idea of waiting based on intervals of
the system tick timer alone. It improves upon it by using a more
accurate algorithm and by taking measures to limit the accumulation of
rounding errors, especially if operating with a very slow system clock.

It also removes the need to have systicks enabled when calling
clock_delay_usec.